Make the 'bin/setup' script useful
This changes the bin/setup script so that it will create a user as part
of the setup process, using an interactive prompt for the name, email
and password.

It also updates the README to recommend using this script, and removes
some caveats now that the fork is a bit more stable and has user 
accounts.