Support 'bot' users in Maslow
This adds a new role to the User model in Maslow for 'bot' users. This
role will be used in future for user accounts which support automated 
tasks like importing needs.

Bot users cannot sign in or request a password reset, and they cannot
be edited through the user interface. Activity feed items for bot users
are denoted with a 'bot' label.