Complete redesign of application
This extends the previous work to redesign the needs view in Maslow to
apply the same look and feel across the whole app.

This introduces the persistent left column throughout the application,
with the navigation items being overridden when viewing the detail of
a need, or the settings panel.

This includes a lot of CSS refactoring, and removes redundant CSS rules
from the previous design.