Add NeedPerformancePoint model
This creates a new model for persisting data points on the performance
of a need response.

A data point is a single value for a single day, and one data point of 
metric type can exist for a response per day.

The NeedPerformancePoint class also contains methods which support
grouping data points by week or month.