Add token authentication for bot users
This adds token authentication through the 'simple_token_authentication'
gem so that bot users can make requests using an API token.

This also changes the way that bot users are restricted from signing in
so that authentication is permitted, but only via token authentication.
We do this by forcing the `valid_password?` method, used by Devise at 
sign-in time, to always return false if the user is a bot.

This adds a primitive JSON format response to the NeedsController, used
primarily for testing.