Add a setup view only accessible when no users exist
This adds a new controller and route for a setup screen which should be displayed when no users exist in the database. It will eventually allow the first user account to be created through the web interface rather than a console.

A check is included to ensure that no users exist. If so, the request is redirected back to the root URL.